backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
numpy>=1.24.0,<2.0.0
# faiss-cpu is optional and may not have wheels on Windows/Python latest.
# If available for your platform, you can uncomment the next line.
# faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
openai==1.43.0
httpx==0.27.0
pypdf==4.3.1
